Rename utils directory to lib
