+++ src/main/java/com/example/rhymedetectorbackend/lib/PWord.java
@@ -1,4 +1,4 @@
-package com.example.rhymedetectorbackend.utils;
+package com.example.rhymedetectorbackend.lib;
 import java.util.ArrayList;
 
 public class PWord {
